fix: publish docker image workflow (#830)
After many tests, I have now discovered that there is currently a
problem with node and arm64 in GitHub workflows. There is the following
issue: https://github.com/nodejs/docker-node/issues/1335. As already
mentioned at the top of the Node.js issue, a possible solution would be
to run it in offline mode, i.e. to cache the packages.

I have now adapted the GitHub workflow file so that it already works.
As you have already correctly noted, I have added QEMU.

Since this problem only affects Node, the backend is not affected and
already has the multi-architecture variant. To give the frontend
multi-arch as well, you would only have to add `linux/arm64` to the
platform, and then it should also work there.

But I still have a question: Why do you not work with the versions in
the workflow file, but with special commits?

---------

Co-authored-by: github-actions[bot] <41898282+github-actions[bot]@users.noreply.github.com>